Implement method for reading full-positions.csv from file
Unfamiliarity with Java meant a lot of time was wasted as I initially didn't understand that the test had to declare which exceptions might be thrown, causing me to
think that there was a problem with my implementation. Also, there doesn't appear to be a preferred method for reading a file into a String.